+++ src/test/java/com/app/attachmentutility/CaptureScreen.java
@@ -1,0 +1,7 @@
+package com.app.attachmentutility;
+
+/**
+ * Created by OS344312 on 8/5/2016.
+ */
+public class CaptureScreen {
+}
