+++ src/test/java/com/app/pom/WebPage2.java
@@ -4,8 +4,12 @@ import com.app.objects.Button;
 import com.app.objects.TextField;
 import com.app.testbase.PageObject;
 import org.openqa.selenium.WebDriver;
+import org.testng.ITestResult;
+import org.testng.TestListenerAdapter;
+import ru.yandex.qatools.allure.annotations.Attachment;
 import ru.yandex.qatools.allure.annotations.Step;
 
+import java.io.File;
 import java.io.IOException;
 
 /**
@@ -35,4 +39,6 @@ public class WebPage2 extends PageObject{
         password.searchKeyword(Password);
         loginBtn.click();
     }
+
+
 }
