+++ src/test/java/com/app/testcases/WebPageTestLocal.java
@@ -5,7 +5,6 @@ import com.app.testbase.WebDriverBase;
 import org.openqa.selenium.OutputType;
 import org.openqa.selenium.TakesScreenshot;
 import org.testng.Assert;
-import org.testng.ITestResult;
 import org.testng.annotations.AfterMethod;
 import org.testng.annotations.BeforeClass;
 import org.testng.annotations.Test;
@@ -18,7 +17,7 @@ import java.io.IOException;
 /**
  * Created by Oscar Garcia on 8/2/2016.
  */
-public class WebPageTest extends WebDriverBase{
+public class WebPageTestLocal extends WebDriverBase{
 
     WebPage _page;
 
@@ -58,28 +57,5 @@ public class WebPageTest extends WebDriverBase{
         Assert.assertTrue(false);
     }
 
-    @AfterMethod
-    @Attachment(value = "Screenshot on failure captured", type = "image/png")
-    public byte[] screenCapture(ITestResult tr){
-
-        if(tr.getStatus()==ITestResult.FAILURE) {
-            String imgPath = "./target/" + tr.getName() + ".png";
-            String path = tr.getName() + ".png";
-            File screenshot = new File(imgPath);
-            try {
-                FileOutputStream screenshotStream = new FileOutputStream(screenshot, false);
-                byte[] bytes = ((TakesScreenshot) _driver)
-                        .getScreenshotAs(OutputType.BYTES);
-                screenshotStream.write(bytes);
-                screenshotStream.close();
-                return bytes;
-            } catch (IOException unableToWriteScreenshot) {
-                System.err.println("Unable to write "
-                        + screenshot.getAbsolutePath());
-                unableToWriteScreenshot.printStackTrace();
-            }
-        }
-        return null;
-    }
 
 }
